Realize the Algorithm Like This Isn't Local
It's probably better to first upsample to a linear grid of high enough
resolution, then downsample to the arbitrary grid. That way, we can get
higher modularity and probably significant speed-ups. Because
interpolating arbitrarily sampled data is non-local, i.e. a far away
point can have large impact and for each interpolated point we'd have to
consider all original points, which is not efficient.